requirements.txt
Flask==3.0.0
openai==0.28.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
python-wordpress-xmlrpc==2.3
flask-cors==4.0.0 